Separate report name and icon configuration from personal details
